Function to impose an overall change impact effect

Function to impose an overall change impact effect

Usage

genOverallchangeData(changecoef.link, model, data, panels = NULL)

Arguments

changecoef.link

Coefficient on the scale of the link function to determine the overall change in the population.

model

A glm or gamMRSea model object.

data

Data frame used to fit model

panels

Character vector denoting the column of data containing the panel structure.

Value

Returns a data frame twice the size of the original input with additional columns for panel id, eventphase and truth. The truth column represents the input data for the first half (pre-event; eventphase==0) and the second half has the change imposed (post-event; eventphase==1).

Author

Lindesay Scott-Hayward, University of St Andrews