Function plot the mean, lower 2.5\% and upper 97.5\% of predicted animal counts before and after the event.

Function plot the mean, lower 2.5\% and upper 97.5\% of predicted animal counts before and after the event.

Usage

# S3 method for class 'preds'
plot(power.object, returndata = FALSE, cellsize)

Arguments

power.object

power analysis object of class gamMRSea.power

returndata

logical stating whether to return the plotting data

cellsize

vector of length two denoting the size of the gridcells (width, height). If unspecified, geom_raster is used for plotting rather than geom_tile.

Author

LAS Scott-Hayward, University of St Andrews

Examples

# Load a power object that was created with a 20\% sitewide decline
# 50 simulations were run
data(nysted.power.oc)

plot.preds(nysted.power.oc)
#> Error in plot.preds(nysted.power.oc): argument "cellsize" is missing, with no default