Function to plot the output from powerSim

Function to plot the output from powerSim

Usage

# S3 method for class 'sigdiff'
plot(
  powerout,
  coordinates,
  tailed = "two",
  error.rate = 0.05,
  adjustment = "none",
  gridcell.dim = NULL
)

Arguments

powerout

power analysis object of class 'gamMRSea.power'

coordinates

nx2 dataframe or matrix of data coordinates

tailed

takes on the value 'one' or 'two' depending on the type of test required

error.rate

number stating the error rate required, default is 0.05 (5\% error rate)

adjustment

character stating the adjustment, if any, for multiple testing. The default value is 'none'. Alternatives are 'sidak' and 'bonferroni'.

gridcell.dim

vector of length 2 denoting the dimensions of the grid cells to be plotted.

Value

Figure showing, for every grid cell, the proportion of simulations that showed a significant difference.

Author

Lindesay Scott-Hayward (University of St Andrews)

Examples

# Load two power objects that were created with a 20% sitewide decline
# 50 simulations were run
data(nysted.power.oc)
data(nysted.power.oc.null)

plot.sigdiff(powerout.nysted.oc, coordinates =
             predictdata[predictdata$eventphase==0,c('x.pos', 'y.pos')],
             tailed='two', error.rate = 0.05, gridcell.dim = c(1000,1000))
#> Loading required package: ggplot2
#> Error: object 'powerout.nysted.oc' not found