Function to estimate the coverage of the overall change impact coefficient.

Function to estimate the coverage of the overall change impact coefficient.

Usage

pval.coverage.null(
  newdat.ind,
  newdat.corr = NULL,
  model,
  nsim,
  powercoefid,
  empdistnull = NULL
)

Arguments

newdat.ind

simulated independent response data

newdat.corr

(default= NULL). simulated correlated response data (if available).

model

glm or gamMRSea model object

nsim

number of simulations in newdat.ind

powercoefid

coefficient id that relates to the eventchange term

empdistnull

if known, provide the empirical distribution for the runs test.

Value

An list object containing:

imppvals

p-values for the event change term for each simulated run

betacis

Confidence intervals for the event term coefficient (only if site-wide change)

empdistnull

Null distribution of runs test statistics.

Author

Lindesay Scott-Hayward, University of St Andrews

Examples

# see the vignette for using this function