Summary function for objects of class gamMRSea.power

Summary function for objects of class gamMRSea.power

Usage

# S3 method for class 'gamMRSea.power'
summary(power.object, null.object = NULL, truebeta = NULL)

Arguments

power.object

An object of class gamMRSea.power

null.object

Object containing information regarding the error rate when no change is present in the data.

truebeta

Value indicating the proportion of data affected by the change. Only required for a site wide change.

Value

printed summary output

If only the power.object is provided, then the summary is reduced or informs the reader of some unknowns.

Author

LAS Scott-Hayward

Examples

# Load two power objects that were created with a 20% sitewide decline
# 50 simulations were run
data(nysted.power.oc)
data(nysted.power.oc.null)

summary(nysted.power.oc, nysted.power.oc.null, truebeta=log(0.8))
#> 
#>  ++++ Summary of Power Analysis Output ++++
#> 
#> Number of power simulations =  50
#> Number of no change simulations =  50 
#> 
#> Power to select 'change' term (eventphase term):
#> 
#>     Under Change (true parameter = -0.2231436) = 70%
#>     Under no change (true parameter = 0) = 6%
#> 
#> Coverage for 'change' coefficient:
#> 
#>     Under model = 94%
#>     Under no change = 94%
#> 
#> Overall Abundance Summary with 95% Confidence Intervals:
#> 
#>        Abundance LowerCI UpperCI
#> Before    665944  546539  795944
#> After     530319  427783  619057
#> 
#> Note: These calculations assume the correct area has been given for each prediction grid cell.